+++ static-deploy.php
@@ -16,6 +16,17 @@ if ( ! defined( 'ABSPATH' ) ) {
 define( 'STATIC_DEPLOY_VERSION', '9.2.1' );
 define( 'STATIC_DEPLOY_PATH', plugin_dir_path( __FILE__ ) );
 
+if ( ! defined( 'STATIC_DEPLOY_DEBUG' ) ) {
+    if (
+        WP_DEBUG
+        || ( defined( 'WP_CLI' ) && WP_CLI::get_config( 'debug' ) )
+    ) {
+        define( 'STATIC_DEPLOY_DEBUG', true );
+    } else {
+        define( 'STATIC_DEPLOY_DEBUG', false );
+    }
+}
+
 if ( file_exists( STATIC_DEPLOY_PATH . 'vendor/autoload.php' ) ) {
     require_once STATIC_DEPLOY_PATH . 'vendor/autoload.php';
 }