+++ views/advanced-options-page.php
@@ -34,7 +34,6 @@ $row = function ( $name ) use ( $options ) {
 
     <table class="widefat striped">
         <tbody>
-            <?php echo $row( 'debugLogging' ); ?>
             <?php echo $row( 'maxLogRows' ); ?>
         </tbody>
     </table>